Implement basic zpool status support
This commit teaches hbsdmon how to monitor a ZFS pool. An alert is
generated only if the pool is in an error/faulted state. No more detail
or context is provided in the alert, only that the pool needs attention.
An example configuration for the node would include:

```
host: "localhost",
method: "ZFS",
pool: "rpool"
```

Note that the host parameter is ignored.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	BlackhawkNest, Inc
issue:		#7