Teach the monitor to name itself
If multiple monitors are deployed in a given infrastructure, it may
become difficult to determine which monitor is alerting. In order to be
able to determine the source of an alert, add a new configuration
parameter (`name`) to name a given monitor instance.

The `name` configuration setting is optional. The default monitor name
is "HardenedBSD Monitor".

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	BlackhawkNest, Inc