Fix heartbeats
The heartbeat mechanism was firing too soon (every second instead of
the default of every six hours). When the application starts, we now
cache the start time as if it's the last heartbeat. This ensures that
the calculation of when do perform the next heartbeat succeeds.

Provide the heartbeat timeout to the main thread's zmq_poll. The main
thread doesn't need a "once every second" poll.

The goal of this thread is to fix spamming Pushover with heartbeat
messages.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	BlackhawkNest, Inc